Rewrite filterConfig.values when material-types are loaded
I am not sure about this approach: it feels slightly naughty, somehow,
to modify the structure of the previously-constant filterConfig
structure, and I am not 100% satisfied that the componentWillUpdate
lifecycle method is the right place to do it. But it works pretty
neatly, and involves minimal code changes.

Fixes STRIPES-287.