Make material type and loan type update properly (STRIPES-355)
Loan types were actually already working correctly (sort of). Part of the confusion
was that the latest black box (v0.16.0) did not include the necessary back-end
changes with support for material type and loan type. There's still something odd
going on, though. THis commit does a lookup of type name from ID that should be
unnecessary.