Material-type sorting uses MaterialTypeId for now
This gives us SOME kind of sorting on material-type, while we try to
figure out how to sort of material-type name -- something which
apparently is not presently supported in the back-end. At least now
items with the same material-type get grouped together, even if the
order of the groups is random.

Partial solution to UIIT-28.